products-agent/pyproject.toml
[project]
name = "strands-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.4"
dependencies = [
    "strands-agents[ollama,openai]>=1.4.0",
    "strands-agents-tools>=0.2.3",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
]
